Combine multiple_targets
