+++ docs/conf.py
@@ -31,8 +31,7 @@ extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.autosummary',
     'sphinx_autodoc_toolbox.collapse',
-    'sphinx_autodoc_toolbox.pydomain_multiple_targets',
-    'sphinx_autodoc_toolbox.autosummary_multiple_targets',
+    'sphinx_autodoc_toolbox.multiple_targets',
     'sphinx_copybutton',
 ]
